+++ CHANGELOG.md
@@ -3,6 +3,11 @@
 ## [Unreleased]
 
 ### Added
+- executeSyncOperation method
+
+## [0.2.0] - 2021-11-15
+
+### Added
 - executeAsyncOperation method
 
 ## [0.1.0] - 2021-11-03